% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gamma.R
\name{compute_gamma}
\alias{compute_gamma}
\title{Group Responsibilities (Z)}
\usage{
compute_gamma(x, y, family, pi, beta, ...)
}
\arguments{
\item{x}{Predictor/design matrix. A numeric matrix of size n x (p + 1), where
the number of rows is equal to the number of observations n, and the number
of columns is equal to the number of covariates p + 1 (for the intercept
term).}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one (i.e. matrix with one column). If family is Binomial, y becomes a numeric
matrix of size n x 2, where the first column corresponds to the successes and
the second the failures.}

\item{family}{A string of characters specifying the distribution of the
finite mixture regression model being fit to the data. Parameter updates
are altered depending on the inputted family.}

\item{pi}{Mixing proportions for each component. Either a numeric vector, or
something coercible to one.}

\item{beta}{Regression parameters for each mixture component (group). A
numeric matrix of size G x (p + 1), where the number of rows is equal to the
number of mixture components G, and the number of columns is equal to the
number of covariates p + 1.}

\item{...}{Additional arguments for computing the group responsibilities
depending on the inputted family.}
}
\value{
A numeric matrix of size n x G, where the number of rows is equal to
the number of observations n, and the number of columns is equal to the
number of mixture components G, representing the group responsibilities for
the given model.
}
\description{
Compute group responsibility matrix for finite mixture regression
distributions. This function is used during model estimation, specifically
within iterations of the MM algorithm.
}
\keyword{internal}
