% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_N_FGMRM.R
\name{compute_N_FGMRM}
\alias{compute_N_FGMRM}
\title{Column Sums of Group Responsibility Matrix}
\usage{
compute_N_FGMRM(gamma_mat)
}
\arguments{
\item{gamma_mat}{Group responsibility matrix. A numeric matrix of size n x G,
where the number of rows is equal to the number of observations n, and the
number of columns is equal to the number of mixture components (groups) G.}
}
\value{
A numeric vector containing the column sums of gamma_mat.
}
\description{
Compute sum of predicted responsibilities along each mixture component
(group) for updating mixing proportions (pi) and sigma estimates for finite
Gaussian mixture regression models. This function is used during model
estimation, specifically within iterations of the MM algorithm.
}
\keyword{internal}
