% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_MM.R
\name{error_check_MM}
\alias{error_check_MM}
\title{Error Check Function for Arguments of MM().}
\usage{
error_check_MM(
  x,
  y,
  G,
  family,
  tol,
  max_iter,
  reps,
  lambda,
  alpha,
  init_parameters,
  verbose,
  penalty,
  information_criteria,
  common_sigma,
  sigma_penalty,
  pi_penalty
)
}
\arguments{
\item{x}{Predictor/design matrix. A numeric matrix of size n x p where the
number of rows is equal to the number of observations n, and the number of
columns is equal to the number of covariates p.}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one (i.e. matrix with one column). If family is Binomial, y becomes a numeric
matrix of size n x 2, where the first column corresponds to the successes and
the second the failures.}

\item{G}{An integer greater than or equal to one representing the
number of mixture components (groups) in a finite mixture regression
model.}

\item{family}{A string of characters specifying the distribution of the
finite mixture regression model being fit to the data. Parameter updates
are altered depending on the inputted family.}

\item{tol}{A non-negative numeric value specifying the stopping criteria for
the MM algorithm (default value is 10e-04). If the difference in value of the
objective function being minimized is within tol in two consecutive
iterations, the algorithm stops.}

\item{max_iter}{An integer greater than or equal to one specifying the
maximum number of iterations ran within the MM algorithm. Default value is
500.}

\item{reps}{An integer greater than or equal to one specifying the
number of times the MM algorithm is repeated on the same initial parameters.
Default value is 1.}

\item{lambda}{A non-negative numeric value (tuning parameter) specifying the
strength of the sparse group lasso (sgl) penalty. Default value is zero (no penalty
applied).}

\item{alpha}{A numeric value between zero (default value) and one inclusive
specifying the weight between the lasso penalty and group lasso penalty being
applied. Alpha = 1 gives the lasso fit and alpha = 0 gives the group lasso
fit.}

\item{init_parameters}{description}

\item{verbose}{A logical value which, if true (default value), allows the
function to print progress updates.}

\item{penalty}{A logical value which, if true (default value), allows the
function to apply the sgl penalty to the regression parameter
updates and objective function within iterations of the MM algorithm.}

\item{common_sigma}{A logical value which, if true (false is the default value)
and family = "gaussian" or gaussian(), estimates the standard deviations as
equivalent across mixture components.}

\item{sigma_penalty}{A logical value which, if true (default value)
and family = "gaussian" or gaussian(), allows a variance-induced penalty to
be applied to the objective function being minimized within the MM algorithm.}

\item{pi_penalty}{A logical value which, if true (default value), allows the
MM algorithm to use estimates for pi in other parameter updates. If false,
all values in the pi vector are replaced with the value one.}
}
\value{
No return value, called for side effects.
}
\description{
Error Check Function for Arguments of MM().
}
\keyword{internal}
