% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMRM.R
\name{FMRM}
\alias{FMRM}
\title{Title}
\usage{
FMRM(
  x,
  y,
  G,
  reps = 1,
  sims = 1,
  tol = 0.001,
  max_iter = 500,
  lambda = NULL,
  lambda_max = NULL,
  n_lambda = 100,
  alpha = seq(0, 1, by = 0.1),
  verbose = TRUE,
  penalty = TRUE,
  random = FALSE,
  n_random_la = 100,
  automatic_stopping = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{parallel}{}
}
\description{
Title
}
