% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_max_compute_FPMRM.R
\name{lambda_max_compute_FPMRM}
\alias{lambda_max_compute_FPMRM}
\title{Title}
\usage{
lambda_max_compute_FPMRM(x, y, z_mat, beta_init)
}
\arguments{
\item{x}{description}

\item{y}{description}

\item{z_mat}{description}

\item{beta_init}{description}
}
\value{
description
}
\description{
Description
}
\keyword{internal}
