% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_max_compute_GLM.R
\name{lambda_max_compute_GLM}
\alias{lambda_max_compute_GLM}
\title{Maximum Lambda (Tuning Parameter)}
\usage{
lambda_max_compute_GLM(x, y, family, z_mat, beta_init)
}
\arguments{
\item{x}{Predictor/design matrix. A numeric matrix of size n x (p + 1), where
the number of rows is equal to the number of observations n, and the number
of columns is equal to the number of covariates p + 1 (for the intercept
term).}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one (i.e. matrix with one column). If family is Binomial, y becomes a numeric
matrix of size n x 2, where the first column corresponds to the successes and
the second the failures.}

\item{family}{A string of characters specifying the distribution of the
finite mixture regression model being fit to the data. Parameter updates
are altered depending on the inputted family.}

\item{z_mat}{A numeric matrix of size n x G, where the number of rows is equal to
the number of observations n, and the number of columns is equal to the
number of mixture components G, representing the group responsibilities for
the given model.}

\item{beta_init}{Regression parameters for each mixture component (group). A
numeric matrix of size G x (p + 1), where the number of rows is equal to the
number of mixture components G, and the number of columns is equal to the
number of covariates p + 1.}
}
\value{
A non-negative numeric value representing the maximum lambda value
to be used for the sgl penalty.
}
\description{
Compute maximum lambda value for sparse group lasso (sgl) penalty applied to
finite mixture regression distribution beta updates in GLM form.
This function is used during model estimation, specifically
in MM_Grid() when initializing the lambda-alpha penalty grid. This computation
differs from the one for finite Gaussian mixture regression distributions.
}
\keyword{internal}
