% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Incomplete Data Log-likelihood for Finite Mixture Regression
Distributions}
\usage{
log_likelihood(x, y, family, pi, beta, ...)
}
\arguments{
\item{x}{Predictor/design matrix. A numeric matrix of size n x (p + 1), where
the number of rows is equal to the number of observations n, and the number
of columns is equal to the number of covariates p + 1 (for the intercept
term).}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one.}

\item{family}{A string of characters specifying the distribution of the
finite mixture regression model being fit to the data. Parameter updates
are altered depending on the inputted family.}

\item{pi}{Mixing proportions for each component. Either a numeric vector, or
something coercible to one.}

\item{beta}{Regression parameters for each mixture component (group). A
numeric matrix of size G x (p + 1), where the number of rows is equal to the
number of mixture components G, and the number of columns is equal to the
number of covariates p + 1 (for the intercept term).}

\item{...}{Additional arguments for computing the log likelihood
depending on the inputted family.}
}
\value{
A numeric scalar representing the incomplete data log-likelihood
for the given model.
}
\description{
Compute incomplete data log-likelihood for finite mixture
regression distributions. This function is used during model estimation,
specifically within iterations of the MM algorithm.
}
\keyword{internal}
