% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM.R
\name{MM}
\alias{MM}
\title{Title}
\usage{
MM(
  x,
  y,
  G,
  reps = 1,
  tol = 0.001,
  max_iter = 500,
  lambda = 0,
  alpha = 0,
  init_pi = NULL,
  init_beta = NULL,
  init_sigma = NULL,
  init_gamma = NULL,
  verbose = TRUE,
  penalty = TRUE
)
}
\arguments{
\item{penalty}{}
}
\description{
Title
}
