% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM_FGMRM.R
\name{MM_FGMRM}
\alias{MM_FGMRM}
\title{Majorization-Minimization Algorithm for Finite Gaussian Mixture Regression
Models}
\usage{
MM_FGMRM(
  x,
  y,
  G,
  tol = 0.001,
  max_iter = 500,
  lambda = 0,
  alpha = 0,
  init_pi = NULL,
  init_beta = NULL,
  init_sigma = NULL,
  init_gamma = NULL,
  verbose = TRUE,
  penalty = TRUE
)
}
\arguments{
\item{x}{Design matrix. A numeric matrix of size n x p where the number of
rows is equal to the number of observations n, and the number of columns is
equal to the number of covariates p.}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one.}

\item{G}{An integer greater than or equal to one representing the
number of mixture components (groups) in a finite Gaussian mixture regression
model.}

\item{tol}{A non-negative numeric value specifying the stopping criteria for
the MM algorithm (default value is 10e-04). If the difference in value of the
objective function being minimized is within tol in two consecutive
iterations, then the algorithm stops.}

\item{max_iter}{An integer greater than or equal to one specifying the
maximum number of iterations ran within the MM algorithm. Default value is
500.}

\item{lambda}{A non-negative numeric value (tuning parameter) specifying the
strength of the sparse group lasso penalty. Default value is zero (no penalty
applied).}

\item{alpha}{A numeric value between zero and one inclusive specifying the
weight between the lasso penalty and group lasso penalty being applied (GS).
Alpha = 1 gives the lasso fit and alpha = 0 gives the group lasso fit (GS).}

\item{init_pi}{A numeric vector containing an (optional) initial estimate for
the mixing proportions of the finite Gaussian mixture model being estimated.
Default value is NULL and if an initial estimate is not provided, one is
initialized using the Mclust function from the mclust package.}

\item{init_beta}{(Optional) Initial estimate for the regression parameters
for each mixture component (group) of the finite Gaussian mixture model being
estimated. Default value is NULL and if an initial estimate is not provided,
one is initialized using the Mclust function from the mclust package. A
numeric matrix of size G x (p + 1), where the number of rows is equal to the
number of mixture components (groups) G, and the number of columns is equal
to the number of covariates p + 1 (for the intercept term).}

\item{init_sigma}{A numeric vector containing an (optional) initial estimate
for the standard deviations of the finite Gaussian mixture model being
estimated. Default value is NULL and if an initial estimate is not provided,
one is initialized using the Mclust function from the mclust package.}

\item{init_gamma}{(Optional) Initial estimate for the group responsibilities
of the finite Gaussian mixture model being estimated. Default value is NULL
and if an initial estimate is not provided, one is initialized using the
Mclust function from the mclust package. A numeric matrix of size n x G,
where the number of rows is equal to the number of observations n, and the
number of columns is equal to the number of mixture components (groups) G.}

\item{verbose}{A logical value which, if true (default value), allows the
function to print progress updates.}

\item{penalty}{A logical value which, if true (default value), allows the
function to apply the sparse group lasso penalty to the regression parameter
updates and objective function within iterations of the MM algorithm.}
}
\value{
A list containing the parameters of the estimated finite Gaussian
mixture regression model (bic, log_likelihood, beta, pi, sigma, z, z_hard,
y_hat, mse, mse_fitted, alpha, lambda).
}
\description{
Applies the Majorization-Minimization Algorithm to the inputted data given
the specified parameters to estimate a finite Gaussian mixture regression
model. Initial estimates for model parameters (pi, beta, sigma, z) are
provided within the function using the Mclust function from the mclust
package, but can be specified in the function call. This function is used
during model estimation.
}
\examples{

set.seed(2025)

# ----Simulate data----
n <- 500   # total samples
p <- 3     # number of covariates
G <- 3     # number of mixture components
rho = 0.2  # correlation

# ----True parameters for 3 clusters----
betas <- matrix(c(
  1,  2, -1,  0.5,   # Component 1
  5, -2,  1,  1.5,   # Component 2
  -3, 0,  2, -1      # Component 3
), nrow = G, byrow = TRUE)
pis <- c(0.4, 0.4, 0.2)
sigmas <- c(3, 1.5, 1)

# ----Generate correlation matrix----
cor_mat <- outer(1:p, 1:p, function(i, j) rho^abs(i - j))
Sigma <- cor_mat

# ----Simulate design matrix X (n × p)----
X <- mvtnorm::rmvnorm(n, mean = rep(0, p), sigma = Sigma)

# ----Generate responsibilities----
z <- rmultinom(n, size = 1, prob = pis)
groups <- apply(z, 2, which.max)

# ----b0 + b1x1 + b2x2 + ... + bkxp----
mu_vec <- rowSums(cbind(1, X) * betas[groups, ])

# ----Simulate response y----
y <- rnorm(n, mean = mu_vec, sd = sigmas[groups])

mod <- MM_FGMRM(X, y, G = 3, lambda = 5, alpha = 1)
}
