% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM_over_lambda_alpha.R
\name{MM_over_lambda_alpha}
\alias{MM_over_lambda_alpha}
\title{Majorization-Minimization Algorithm over Lambda-Alpha Vectors}
\usage{
MM_over_lambda_alpha(
  g,
  x,
  y,
  reps = 1,
  tol = 0.001,
  max_iter = 500,
  lambda = NULL,
  lambda_max = NULL,
  n_lambda = 100,
  alpha = seq(0, 1, by = 0.1),
  verbose = TRUE,
  penalty = TRUE,
  random = FALSE,
  n_random_la = 100,
  parallel = TRUE
)
}
\arguments{
\item{x}{Design matrix. A numeric matrix of size n x p where the number of
rows is equal to the number of observations n, and the number of columns is
equal to the number of covariates p.}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one.}

\item{n_random_la}{}
}
\description{
Majorization-Minimization Algorithm over Lambda-Alpha Vectors
}
