% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_FGMRM.R
\name{objective_function_FGMRM}
\alias{objective_function_FGMRM}
\title{Objective Function for a Finite Gaussian Mixture Regression Distribution}
\usage{
objective_function_FGMRM(ll, pen)
}
\arguments{
\item{ll}{A numeric scalar representing the log-likelihood of the finite
Gaussian mixture regression model.}

\item{pen}{A numeric scalar representing the sparse group lasso penalty
being applied to the log-likelihood.}
}
\value{
A numeric scalar representing the negative penalized log-likelihood
objective function used for minimization.
}
\description{
Computes the negative penalized log-likelihood objective function for a
finite Gaussian mixture regression distribution. This function is used during
model estimation, specifically within iterations of the MM algorithm.
}
\keyword{internal}
