% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function_FPMRM.R
\name{objective_function_FPMRM}
\alias{objective_function_FPMRM}
\title{Objective Function for a Finite Poisson Mixture Regression Distribution}
\usage{
objective_function_FPMRM(ll, pen, n)
}
\arguments{
\item{ll}{A numeric scalar representing the log-likelihood of the finite
Poisson mixture regression model.}

\item{pen}{A numeric scalar representing the sparse group lasso (sgl) penalty
being applied to the log-likelihood.}

\item{n}{A numeric value representing the number of observations in the data.}
}
\value{
A numeric scalar representing the negative penalized log-likelihood
objective function used for minimization.
}
\description{
Computes the negative penalized log-likelihood objective function for a
finite Poisson mixture regression distribution. This function is used during
model estimation, specifically within iterations of the MM algorithm.
}
\keyword{internal}
