% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_MM.R
\name{penalty_MM}
\alias{penalty_MM}
\title{Calculate sparse group lasso penalty}
\usage{
penalty_MM(lambda, alpha, beta, G)
}
\arguments{
\item{lambda}{Tuning parameter}

\item{alpha}{ADD HERE}

\item{beta}{Regression parameters for each group of finite Gaussian mixture
regression model. A numeric matrix of size G x (p + 1), where the number of
rows is equal to the number of groups G, and the number of columns is
equal to the number of covariates p + 1}

\item{G}{Number of groups in finite Gaussian mixture regression model}
}
\value{
A single numeric value
}
\description{
ADD DESCRIPTION
}
