% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_update.R
\name{pi_update}
\alias{pi_update}
\title{Update mixing proportions for finite Gaussian mixture regression model within
MM Algorithm}
\usage{
pi_update(n, gamma_mat)
}
\arguments{
\item{gamma_mat}{}
}
\description{
Update mixing proportions for finite Gaussian mixture regression model within
MM Algorithm
}
