% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FPMRM.R
\name{plot.FPMRM}
\alias{plot.FPMRM}
\title{Plot Method for a Finite Poisson Mixture Regression Model of class "FPMRM"}
\usage{
\method{plot}{FPMRM}(x, ...)
}
\arguments{
\item{x}{An object of class "FPMRM", the result of calling FMRM() or
MM_Grid() with family = poisson().}

\item{...}{Additional arguments for plotting (currently unused).}
}
\value{
A list of three ggplot objects: lambdas vs. bics for all alpha
values, lambdas vs. regression coefficients, and lambdas vs. group norms for
all models with the same alpha as the optimal alpha.
}
\description{
This function creates plots for finite Poisson mixture regression models of
class "FPMRM". It generates three plots: lambdas vs. bics for all alpha
values, lambdas vs. regression coefficients, and lambdas vs. group norms for
all models with the same alpha as the optimal alpha.
}
\examples{
\dontshow{if (rlang::is_installed("mvtnorm")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

set.seed(2025)

# ----Simulate data----
n <- 500  # total samples
p <- 6    # number of covariates
G <- 3    # number of mixture components
rho <- 0.2  # correlation

# ----True parameters for 3 clusters----
betas <- matrix(c(
  1,  2, -1,  0.5, 0, 0, 0,  # component 1
  5, -2,  1,  0,   0, 0, 0,  # component 2
  -3,  0,  2,  0,   0, 0, 0   # component 3
  ), nrow = G, byrow = TRUE) / 2
  pis <- c(0.4, 0.4, 0.2)

# ----Generate correlation matrix----
cor_mat <- outer(1:p, 1:p, function(i, j) rho^abs(i - j))
Sigma <- cor_mat

# ----Simulate design matrix X (n x p)----
X <- mvtnorm::rmvnorm(n, mean = rep(0, p), sigma = Sigma)

# ----Generate responsibilities----
z <- rmultinom(n, size = 1, prob = pis)
groups <- apply(z, 2, which.max)

# ----b0 + b1x1 + b2x2 + ... + bkxp (log-linear predictor)----
eta_vec <- rowSums(cbind(1, X) * betas[groups, ])

# ----Apply inverse link (exp) to get Poisson means----
mu_vec <- exp(eta_vec)

# ----Simulate response y (count data)----
y <- rpois(n, lambda = mu_vec)

mod <- FMRM(x = X, y = y, G = 6, family = poisson(), verbose = FALSE)

# ----Call plot----
plots <- plot(mod)

# ----Display plots----
plots[[1]] # ----lambdas vs. bics----
plots[[2]] # ----lambdas vs. regression coefficients----
plots[[3]] # ----lambdas vs. group norms----
\dontshow{\}) # examplesIf}
}
