% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FMRM.R
\name{print.FMRM}
\alias{print.FMRM}
\title{Print Method for a Finite Mixture Regression Model of Class FMRM}
\usage{
\method{print}{FMRM}(x, ...)
}
\arguments{
\item{x}{An object of class FMRM, the result of calling FMRM() or
MM_Grid().}

\item{...}{Additional arguments for printing (currently unused).}
}
\value{
No return value, called for side effects.
}
\description{
This function prints the elements of finite mixture regression
models of class FMRM. It displays the number of mixture components,
optimal lambda-alpha, log-likelihood, information criteria, mean-squared-error, and
parameters of the model.
}
\examples{
\dontshow{if (rlang::is_installed("mvtnorm")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

set.seed(2025)

# ----Simulate data----
n <- 500   # total samples
p <- 6     # number of covariates
G <- 3     # number of mixture components
rho = 0.2  # correlation

# ----True parameters for 3 clusters----
betas <- matrix(c(
  1,  2, -1,  0.5, 0, 0, 0,  # component 1
  5, -2,  1,  0, 0, 0, 0,  # component 2
  -3, 0,  2, 0, 0, 0, 0     # component 3
), nrow = G, byrow = TRUE)
pis <- c(0.4, 0.4, 0.2)
sigmas <- c(3, 1.5, 1)/2

# ----Generate correlation matrix----
cor_mat <- outer(1:p, 1:p, function(i, j) rho^abs(i - j))
Sigma <- cor_mat

# ----Simulate design matrix X (n × p)----
X <- mvtnorm::rmvnorm(n, mean = rep(0, p), sigma = Sigma)

# ----Generate responsibilities----
z <- rmultinom(n, size = 1, prob = pis)
groups <- apply(z, 2, which.max)

# ----b0 + b1x1 + b2x2 + ... + bkxp----
mu_vec <- rowSums(cbind(1, X) * betas[groups, ])

# ----Simulate response y----
y <- rnorm(n, mean = mu_vec, sd = sigmas[groups])

mod <- FMRM(x = X, y = y, G = 3, family = gaussian(), parallel = TRUE, verbose = FALSE)

# ----Print model----
print(mod)
\dontshow{\}) # examplesIf}
}
