% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_FMRM.R
\name{print_model_FMRM}
\alias{print_model_FMRM}
\title{Title}
\usage{
print_model_FMRM(
  selected_parameters,
  selected_compartment,
  family,
  information_criteria
)
}
\arguments{
\item{selected_parameters}{description}

\item{selected_compartment}{description}

\item{family}{description}

\item{information_criteria}{A string of characters specifying the
information criteria for model selection purposes. The model that minimizes the
information criteria over all group counts and lambda-alpha pairs will be selected.
Current accepted types include BIC ("bic") (default value), EBIC ("ebic"),
and AIC ("aic").}
}
\value{
No return value, called for side effects
}
\description{
Title
}
\keyword{internal}
