% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model_MM_Grid.R
\name{print_model_MM_Grid}
\alias{print_model_MM_Grid}
\title{Printing Helper Function for MM_Grid()}
\usage{
print_model_MM_Grid(chosen_parameters, g, family, information_criteria)
}
\arguments{
\item{chosen_parameters}{parameters for finite mixture regression model,
depends on family.}

\item{g}{An integer greater than or equal to one representing the
number of mixture components (groups) in a finite mixture regression
model.}

\item{family}{A string of characters specifying the distribution of the
finite mixture regression model being fit to the data. Parameter updates
are altered depending on the inputted family.}

\item{information_criteria}{A string of characters specifying the
information criteria for model selection purposes. The model that minimizes the
information criteria over all group counts and lambda-alpha pairs will be selected.
Current accepted types include BIC ("bic") (default value), EBIC ("ebic"),
and AIC ("aic").}
}
\value{
No return value, called for side effects
}
\description{
Printing Helper Function for MM_Grid()
}
\keyword{internal}
