% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_MM.R
\name{simulation_MM}
\alias{simulation_MM}
\title{Title}
\usage{
simulation_MM(
  i,
  x,
  y,
  G,
  reps = 1,
  tol = 1e-07,
  max_iter = 500,
  lambda = NULL,
  lambda_max = NULL,
  n_lambda = 100,
  alpha = seq(0, 1, by = 0.1),
  verbose = TRUE,
  penalty = TRUE,
  random = FALSE,
  n_random_la = 100,
  automatic_stopping = FALSE
)
}
\arguments{
\item{automatic_stopping}{}
}
\description{
Title
}
