% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FBMRM.R
\name{summary.FBMRM}
\alias{summary.FBMRM}
\title{Summary Method for a Finite Binomial Mixture Regression Model of class
"FBMRM"}
\usage{
\method{summary}{FBMRM}(object, ...)
}
\arguments{
\item{object}{An object of class "FBMRM", the result of calling FMRM() or
MM_Grid() with family = binomial().}

\item{...}{Additional arguments for summary (currently unused).}
}
\value{
No return value, called for side effects.
}
\description{
This function displays a summary for finite Binomial mixture regression
models of class "FBMRM". It displays the number of mixture components,
optimal lambda-alpha, log-likelihood, information criteria, mean-squared-error, and
parameters (pi, beta) of the model.
}
\examples{
\dontshow{if (rlang::is_installed("mvtnorm")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

set.seed(2025)

# ----Simulate data----
n <- 500  # total samples
p <- 6    # number of covariates
G <- 3    # number of mixture components
rho <- 0.2  # correlation
m <- 10   # number of trials (binomial size)

# ----True parameters for 3 clusters----
betas <- matrix(c(
  1,  2, -1,  0.5, 0, 0, 0,  # component 1
  5, -2,  1,  0,   0, 0, 0,  # component 2
  -3,  0,  2,  0,   0, 0, 0   # component 3
), nrow = G, byrow = TRUE) / 2
pis <- c(0.4, 0.4, 0.2)

# ----Generate correlation matrix----
cor_mat <- outer(1:p, 1:p, function(i, j) rho^abs(i - j))
Sigma <- cor_mat

# ----Simulate design matrix X (n x p)----
X <- mvtnorm::rmvnorm(n, mean = rep(0, p), sigma = Sigma)

# ----Generate responsibilities----
z <- rmultinom(n, size = 1, prob = pis)
groups <- apply(z, 2, which.max)

# ----b0 + b1x1 + b2x2 + ... + bkxp (logit-linear predictor)----
eta_vec <- rowSums(cbind(1, X) * betas[groups, ])

# ----Apply inverse link (logistic) to get probabilities----
prob_vec <- plogis(eta_vec)

# ----Simulate response y (count of successes out of m trials)----
y <- rbinom(n, size = m, prob = prob_vec)

mod <- FMRM(x = X, y = y, G = 4, family = binomial(), verbose = FALSE)

# ----Display summary----
summary(mod)
\dontshow{\}) # examplesIf}
}
