% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FGMRM.R
\name{summary.FGMRM}
\alias{summary.FGMRM}
\title{Title}
\usage{
\method{summary}{FGMRM}(object, ...)
}
\arguments{
\item{object}{An object of class "FGMRM", the result of calling FGMRM() or
MM_Grid_FGMRM().}

\item{...}{Additional arguments for summary (currently unused).}
}
\description{
Title
}
