% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gamma.R
\name{compute_gamma}
\alias{compute_gamma}
\title{Compute group responsibility matrix for a finite Gaussian mixture regression
distribution}
\usage{
compute_gamma(x, y, pi, beta, sigma)
}
\arguments{
\item{sigma}{}
}
\description{
Compute group responsibility matrix for a finite Gaussian mixture regression
distribution
}
