% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_N.R
\name{compute_N}
\alias{compute_N}
\title{Compute sum of predicted responsibilities across each group for updating
mixing proportions and sigma estimates for finite Gaussian mixture regression
model within MM Algorithm}
\usage{
compute_N(gamma_mat)
}
\arguments{
\item{gamma_mat}{Group responsibility matrix. A numeric matrix of size n x G,
where the number of rows is equal to the number of observations n, and the
number of columns is equal to the number of groups G}
}
\value{
A numeric vector
}
\description{
ADD DESCRIPTION
}
