% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_V.R
\name{compute_V}
\alias{compute_V}
\title{Compute sparse group lasso penalty majorization term for updating
regression parameters for finite Gaussian mixture regression model within MM
Algorithm}
\usage{
compute_V(G, beta, alpha)
}
\arguments{
\item{alpha}{}
}
\description{
Compute sparse group lasso penalty majorization term for updating
regression parameters for finite Gaussian mixture regression model within MM
Algorithm
}
