% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_V.R
\name{compute_V}
\alias{compute_V}
\title{Sparse Group Lasso Penalty Majorization Matrix}
\usage{
compute_V(G, beta, alpha)
}
\arguments{
\item{G}{An integer greater than or equal to one representing the
number of mixture components (groups) in a finite Gaussian mixture regression
model.}

\item{beta}{Regression parameters for each mixture component (group). A
numeric matrix of size G x (p + 1), where the number of rows is equal to the
number of mixture components (groups) G, and the number of columns is equal
to the number of covariates p + 1 (for the intercept term).}

\item{alpha}{A numeric value between zero and one inclusive specifying the
weight between the lasso penalty and group lasso penalty being applied (GS).
Alpha = 1 gives the lasso fit and alpha = 0 gives the group lasso fit (GS).}
}
\value{
A numeric matrix of size G x (p + 1), where the number of
rows is equal to the number of mixture components (groups) G, and the number
of columns is equal to the number of covariates p + 1 (for the intercept
term), representing the sgl penalty majorization matrix.
}
\description{
Compute sparse group lasso (sgl) penalty majorization matrix for application
when updating regression parameters for finite Gaussian mixture regression
models. This function is used during model estimation, specifically within
iterations of the MM algorithm.
}
\keyword{internal}
