% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMRM.R
\name{FMRM}
\alias{FMRM}
\title{Finite Mixture Regression Model Fitting}
\usage{
FMRM(
  x,
  y,
  G,
  reps = 1,
  sims = 1,
  tol = 0.001,
  max_iter = 500,
  lambda = NULL,
  lambda_max = NULL,
  n_lambda = 100,
  alpha = seq(0, 1, by = 0.1),
  verbose = TRUE,
  penalty = TRUE,
  random = FALSE,
  n_random_la = 100,
  automatic_stopping = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{x}{Design matrix. A numeric matrix of size n x p where the number of
rows is equal to the number of observations n, and the number of columns is
equal to the number of covariates p.}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one.}

\item{G}{An integer ADD HERE}

\item{reps}{An integer greater than or equal to one specifying the number of
random initializations ran within the MM algorithm. Default value is 1.}

\item{sims}{ADD HERE}

\item{tol}{A non-negative numeric value specifying the stopping criteria for
the MM algorithm. If the difference in value of the objective function being
minimized is within tol in two consecutive iterations, then the algorithm
stops.}

\item{max_iter}{An integer greater than or equal to one specifying the
maximum number of iterations ran within the MM algorithm. Default value is
500.}

\item{lambda}{ADD HERE}

\item{lambda_max}{A numeric value ADD HERE}

\item{n_lambda}{An integer ADD HERE}

\item{alpha}{A numeric vector ADD HERE}

\item{verbose}{A logical value which, if true (default value), prints
progress updates within the function.}

\item{penalty}{A logical value which, if true (default value), applies the
sparse group lasso penalty to the regression parameter updates and objective
function within iterations of the MM algorithm.}

\item{random}{A logical value ADD HERE}

\item{n_random_la}{An integer ADD HERE}

\item{automatic_stopping}{A logical value ADD HERE}

\item{parallel}{A logical value ADD HERE}
}
\value{
ADD HERE
}
\description{
ADD HERE
}
