% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute incomplete data log-likelihood for a finite Gaussian mixture
regression distribution}
\usage{
log_likelihood(x, y, pi, beta, sigma)
}
\arguments{
\item{sigma}{}
}
\description{
Compute incomplete data log-likelihood for a finite Gaussian mixture
regression distribution
}
