% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute incomplete data log-likelihood for a finite Gaussian mixture
regression distribution}
\usage{
log_likelihood(x, y, pi, beta, sigma)
}
\arguments{
\item{x}{Design matrix. A numeric matrix of size n x p, where the number of
rows is equal to the number of observations n, and the number of columns is
equal to the number of covariates p}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one}

\item{pi}{Mixing proportions for each group of finite Gaussian mixture
regression model. Either a numeric vector, or something coercible to one}

\item{beta}{Regression parameters for each group of finite Gaussian mixture
regression model. A numeric matrix of size G x (p + 1), where the number of
rows is equal to the number of groups G, and the number of columns is
equal to the number of covariates p + 1}

\item{sigma}{Standard deviation for each group of finite Gaussian mixture
regression model. Either a numeric vector, or something coercible to one}
}
\value{
A single numeric value
}
\description{
ADD DESCRIPTION
}
