% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute incomplete data log-likelihood for a finite Gaussian mixture
regression distribution}
\usage{
log_likelihood(x, y, pi, beta, sigma)
}
\arguments{
\item{sigma}{Standard deviation for each group of finite Gaussian mixture
regression model. Either a numeric vector, or something coercible to one}
}
\value{
A single numeric value
}
\description{
Compute incomplete data log-likelihood for a finite Gaussian mixture
regression distribution
}
