% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM.R
\name{MM}
\alias{MM}
\title{Majorization-Minimization Algorithm}
\usage{
MM(
  x,
  y,
  G,
  reps = 1,
  tol = 0.001,
  max_iter = 500,
  lambda = 0,
  alpha = 0,
  init_pi = NULL,
  init_beta = NULL,
  init_sigma = NULL,
  init_gamma = NULL,
  verbose = TRUE,
  penalty = TRUE
)
}
\arguments{
\item{x}{Design matrix. A numeric matrix of size n x p where the number of
rows is equal to the number of observations n, and the number of columns is
equal to the number of covariates p.}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one.}

\item{G}{An integer greater than or equal to one representing the
number of mixture components (groups) in a finite Gaussian mixture regression
model.}

\item{lambda}{A non-negative numeric value specifying the strength of the
sparse group lasso penalty. Also known as the tuning parameter. Default value
is 0 (no penalty applied).}

\item{alpha}{A numeric value between zero and one inclusive specifying the
weight between the lasso penalty and group lasso penalty being applied (GS).
Alpha = 1 gives the lasso fit and alpha = 0 (default value) gives the group
lasso fit (GS).}

\item{penalty}{}
}
\description{
Majorization-Minimization Algorithm
}
