% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_fun_MM.R
\name{objective_fun_MM}
\alias{objective_fun_MM}
\title{Calculate objective function for a finite Gaussian mixture regression
model}
\usage{
objective_fun_MM(ll, pen)
}
\arguments{
\item{ll}{Log-likelihood}

\item{pen}{Penalty term}
}
\value{
A single numeric value
}
\description{
ADD DESCRIPTION
}
