% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FGMRM.R
\name{plot.FGMRM}
\alias{plot.FGMRM}
\title{Title}
\usage{
\method{plot}{FGMRM}(mod, ...)
}
\arguments{
\item{...}{Additional arguments for plotting (currently unused)}
}
\value{
A list of three ggplot objects: lambdas vs. bics, lambdas vs.
regression coefficients, and lambdas vs. group norms for all models with the
same alpha as the optimal alpha.
}
\description{
Title
}
