% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.FGMRM.R
\name{plot.FGMRM}
\alias{plot.FGMRM}
\title{Plot Method for a Finite Gaussian Mixture Regression Model of class "FGMRM"}
\usage{
\method{plot}{FGMRM}(x, ...)
}
\arguments{
\item{x}{An object of class "FGMRM", the result of calling FGMRM() or
MM_Grid_FGMRM().}

\item{...}{Additional arguments for plotting (currently unused).}
}
\value{
A list of three ggplot objects: lambdas vs. bics, lambdas vs.
regression coefficients, and lambdas vs. group norms for all models with the
same alpha as the optimal alpha.
}
\description{
This function creates plots for finite Gaussian mixture regression models of
class "FGMRM". It generates three plots: lambdas vs. bics, lambdas vs.
regression coefficients, and lambdas vs. group norms for all models with the
same alpha as the optimal alpha.
}
\examples{

set.seed(2025)

# ----Simulate data----
n <- 500   # total samples
p <- 3     # number of covariates
G <- 3     # number of mixture components
rho = 0.2  # correlation

# ----True parameters for 3 clusters----
betas <- matrix(c(
  1,  2, -1,  0.5,   # Component 1
  5, -2,  1,  1.5,   # Component 2
  -3, 0,  2, -1      # Component 3
), nrow = G, byrow = TRUE)
pis <- c(0.4, 0.4, 0.2)
sigmas <- c(3, 1.5, 1)

# ----Generate correlation matrix----
cor_mat <- outer(1:p, 1:p, function(i, j) rho^abs(i - j))
Sigma <- cor_mat

# ----Simulate design matrix X (n × p)----
X <- mvtnorm::rmvnorm(n, mean = rep(0, p), sigma = Sigma)

# ----Generate responsibilities----
z <- rmultinom(n, size = 1, prob = pis)
groups <- apply(z, 2, which.max)

# ----b0 + b1x1 + b2x2 + ... + bkxp----
mu_vec <- rowSums(cbind(1, X) * betas[groups, ])

# ----Simulate response y----
y <- rnorm(n, mean = mu_vec, sd = sigmas[groups])

mod <- FGMRM(x = X, y = y, G = 6, verbose = FALSE)

plots <- plot(mod)

# ----Display plots----
plots[[1]] # ----lambdas vs. bics----
plots[[2]] # ----lambdas vs. regression coefficients----
plots[[3]] # ----lambdas vs. group norms----
}
