% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{plot2}
\alias{plot2}
\title{Plot Covariate of X Against Y With Group Assignments}
\usage{
plot2(mod, x, y, covariate)
}
\arguments{
\item{mod}{An object of class "FGMRM", the result of calling FGMRM() or
MM_Grid_FGMRM().}

\item{x}{Design matrix. A numeric matrix of size n x p where the number of
rows is equal to the number of observations n, and the number of columns is
equal to the number of covariates p.}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one.}

\item{covariate}{A numeric value specifying the covariate of x to be plotted.}
}
\value{
A ggplot object containing a plot of the specified covariate of x
against y, with the group assignments highlighted in colour.
}
\description{
This function creates a plot for finite Gaussian mixture regression models of
class "FGMRM". It plots the specified covariate of x against y, with the
group assignments highlighted in colour.
}
\examples{

# Simulate data
set.seed(123)

n <- 500
G <- 3
p <- 10
rho = 0.2

# ----true parameters for 3 clusters----
sigma_squared_true <- c(3, 1.5, 1)
pi_true <- c(0.4, 0.4, 0.2)
beta_true <- matrix(c(
-1, -3.22, 0, 0, 0, 0, 0.583, 0, 5.17, 0, 0,
1, 0, 0, 0, 0, 0, -4.56, 0.514, -2.98, 0, 0,
3, 0, 0, 0, 3.11, 0, 0, 0, -3.11, 0, 0
), nrow = G, byrow = TRUE)

# ----generate correlation matrix----
cor_mat <- outer(1:p, 1:p, function(i, j) rho^abs(i - j))
Sigma <- cor_mat

# ----simulate each group----
x <- mvtnorm::rmvnorm(n, mean = rep(0, p), sigma = Sigma)

# ----generate responsibilities----
z <- rmultinom(n, size = 1, prob = pi_true)
groups <- apply(z, 2, which.max)

# ----b0 + b1x1 + b2x2 + ... + bkxk----
mu_vec <- rowSums(cbind(1, x) * beta_true[groups, ])

# ----simulate response y----
y <- rnorm(n, mean = mu_vec, sd = sqrt(sigma_squared_true[groups]))

mod <- FGMRM(x, y, G = 6, verbose = FALSE)

plot <- plot2(mod, x, y, 1)

# ----display plot----
plot
}
