% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_penalty_FGMRM.R
\name{sigma_penalty_FGMRM}
\alias{sigma_penalty_FGMRM}
\title{Title}
\usage{
sigma_penalty_FGMRM(sigma, S_x, a_n)
}
\arguments{
\item{sigma}{description}

\item{S_x}{description}

\item{a_n}{description}
}
\value{
description
}
\description{
Description
}
\keyword{internal}
