% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_MM.R
\name{simulation_MM}
\alias{simulation_MM}
\title{Majorization-Minimization Algorithm over Mixing Compartments}
\usage{
simulation_MM(
  i,
  x,
  y,
  G,
  reps = 1,
  tol = 0.001,
  max_iter = 500,
  lambda = NULL,
  lambda_max = NULL,
  n_lambda = 100,
  alpha = seq(0, 1, by = 0.1),
  verbose = TRUE,
  penalty = TRUE,
  random = FALSE,
  n_random_la = 100,
  automatic_stopping = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{i}{ADD HERE}

\item{x}{Design matrix. A numeric matrix of size n x p where the number of
rows is equal to the number of observations n, and the number of columns is
equal to the number of covariates p.}

\item{y}{Response vector. Either a numeric vector, or something coercible to
one.}

\item{G}{An integer greater than or equal to two specifying the maximum
number of mixture components (groups) in the estimated model that the
function will attempt to fit the data to.}

\item{reps}{An integer greater than or equal to one specifying the number of
random initializations ran within the MM algorithm. Default value is one.}

\item{tol}{A non-negative numeric value specifying the stopping criteria for
the MM algorithm (default value is 10e-04). If the difference in value of the
objective function being minimized is within tol in two consecutive
iterations, then the algorithm stops.}

\item{max_iter}{An integer greater than or equal to one specifying the
maximum number of iterations ran within the MM algorithm. Default value is
500.}

\item{lambda}{A list of length G of numeric vectors containing non-negative
tuning parameters specifying various strengths of the sparse group lasso
penalty. Finite Gaussian mixture models will be estimated using each lambda
value. Default value is NULL as function will initialize lambdas for each
group count from 2 to G using an algorithm.}

\item{lambda_max}{A non-negative numeric value specifying the maximum lambda
value (tuning parameter) used in creation of each lambda vector. Default
value is NULL as function will initialize lambda_max for each group count
from 2 to G using an algorithm.}

\item{n_lambda}{An integer ADD HERE}

\item{alpha}{A numeric vector ADD HERE}

\item{verbose}{A logical value which, if true (default value), allows the
function to print progress updates.}

\item{penalty}{A logical value which, if true (default value), allows the
function to apply the sparse group lasso penalty to the regression parameter
updates and objective function within iterations of the MM algorithm.}

\item{random}{A logical value which, if true (false is the default value),
allows the function to take a random sample of size n_random_la from the
lambda-alpha pairs and run the MM algorithm over the reduced grid.}

\item{n_random_la}{An integer ADD HERE}

\item{automatic_stopping}{A logical value ADD HERE}

\item{parallel}{A logical value ADD HERE}
}
\value{
ADD HERE
}
\description{
ADD HERE
}
\keyword{internal}
